+++ tests/conftest.py
@@ -1,4 +1,4 @@
-# Conftest file for sharing fixtures 
+# Conftest file for sharing fixtures
 # Reference: https://docs.pytest.org/en/latest/reference/fixtures.html#conftest-py-sharing-fixtures-across-multiple-files
 
 import os
@@ -7,6 +7,7 @@ import pytest
 from autosubmitconfigparser.config.basicconfig import BasicConfig
 from autosubmit_api.app import create_app
 from autosubmit_api.config.basicConfig import APIBasicConfig
+from autosubmit_api import config
 from tests.custom_utils import custom_return_value
 
 FAKE_EXP_DIR = "./tests/experiments/"
@@ -15,20 +16,14 @@ FAKE_EXP_DIR = "./tests/experiments/"
 #### FIXTURES ####
 @pytest.fixture(autouse=True)
 def fixture_disable_protection(monkeypatch: pytest.MonkeyPatch):
+    monkeypatch.setattr(config, "PROTECTION_LEVEL", "NONE")
     monkeypatch.setenv("PROTECTION_LEVEL", "NONE")
+
 
 @pytest.fixture
 def fixture_mock_basic_config(monkeypatch: pytest.MonkeyPatch):
-    # Patch APIBasicConfig parent BasicConfig
-    monkeypatch.setattr(BasicConfig, "read", custom_return_value(None))
-    monkeypatch.setattr(APIBasicConfig, "read", custom_return_value(None))
-    monkeypatch.setattr(BasicConfig, "LOCAL_ROOT_DIR", FAKE_EXP_DIR)
-    monkeypatch.setattr(BasicConfig, "DB_DIR", FAKE_EXP_DIR)
-    monkeypatch.setattr(BasicConfig, "DB_FILE", "autosubmit.db")
-    monkeypatch.setattr(
-        BasicConfig, "DB_PATH", os.path.join(FAKE_EXP_DIR, "autosubmit.db")
-    )
-    monkeypatch.setattr(BasicConfig, "AS_TIMES_DB", "as_times.db")
+    # Get APIBasicConfig from file
+    monkeypatch.setenv("AUTOSUBMIT_CONFIGURATION", os.path.join(FAKE_EXP_DIR, ".autosubmitrc"))
     yield APIBasicConfig