+++ tests/test_endpoints_v3.py
@@ -89,6 +89,7 @@ class TestExpInfo:
         expid = "a003"
         response = fixture_client.get(self.endpoint.format(expid=expid))
         resp_obj: dict = response.get_json()
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert resp_obj["expid"] == expid
         assert resp_obj["total_jobs"] == 8
@@ -97,10 +98,11 @@ class TestExpInfo:
         expid = "a3tb"
         response = fixture_client.get(self.endpoint.format(expid=expid))
         resp_obj: dict = response.get_json()
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert resp_obj["expid"] == expid
         assert resp_obj["total_jobs"] == 55
-        assert resp_obj["completed_jobs"] == 24
+        assert resp_obj["completed_jobs"] == 28
 
 
 class TestPerformance:
@@ -113,12 +115,14 @@ class TestPerformance:
         expid = "a007"
         response = fixture_client.get(self.endpoint.format(expid=expid))
         resp_obj: dict = response.get_json()
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert resp_obj["Parallelization"] == 8
 
         expid = "a3tb"
         response = fixture_client.get(self.endpoint.format(expid=expid))
         resp_obj: dict = response.get_json()
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert resp_obj["Parallelization"] == 768
 
@@ -129,6 +133,7 @@ class TestPerformance:
         expid = "a003"
         response = fixture_client.get(self.endpoint.format(expid=expid))
         resp_obj: dict = response.get_json()
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert resp_obj["Parallelization"] == 16
 
@@ -145,6 +150,7 @@ class TestTree:
         )
         resp_obj: dict = response.get_json()
 
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert resp_obj["total"] == 8
         assert resp_obj["total"] == len(resp_obj["jobs"])
@@ -160,6 +166,7 @@ class TestTree:
         )
         resp_obj: dict = response.get_json()
 
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert resp_obj["total"] == 55
         assert resp_obj["total"] == len(resp_obj["jobs"])
@@ -179,6 +186,7 @@ class TestRunsList:
         response = fixture_client.get(self.endpoint.format(expid=expid))
         resp_obj: dict = response.get_json()
 
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert isinstance(resp_obj["runs"], list)
 
@@ -192,6 +200,7 @@ class TestRunDetail:
         response = fixture_client.get(self.endpoint.format(expid=expid, runId=2))
         resp_obj: dict = response.get_json()
 
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert resp_obj["total"] == 8
 
@@ -204,6 +213,7 @@ class TestQuick:
         response = fixture_client.get(self.endpoint.format(expid=expid))
         resp_obj: dict = response.get_json()
 
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert resp_obj["total"] == len(resp_obj["tree_view"])
         assert resp_obj["total"] == len(resp_obj["view_data"])
@@ -220,7 +230,7 @@ class TestGraph:
             query_string={"loggedUser": random_user},
         )
         resp_obj: dict = response.get_json()
-
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert resp_obj["total_jobs"] == len(resp_obj["nodes"])
 
@@ -234,7 +244,7 @@ class TestGraph:
             query_string={"loggedUser": random_user},
         )
         resp_obj: dict = response.get_json()
-
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert resp_obj["total_jobs"] == len(resp_obj["nodes"])
 
@@ -246,7 +256,7 @@ class TestGraph:
             query_string={"loggedUser": random_user},
         )
         resp_obj: dict = response.get_json()
-
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert resp_obj["total_jobs"] == len(resp_obj["nodes"])
 
@@ -258,7 +268,45 @@ class TestGraph:
             query_string={"loggedUser": random_user},
         )
         resp_obj: dict = response.get_json()
+        assert resp_obj["error_message"] == ""
+        assert resp_obj["error"] == False
+        assert resp_obj["total_jobs"] == len(resp_obj["nodes"])
 
+    def test_graph_standard_none_retro3(self, fixture_client: FlaskClient):
+        expid = "a3tb"
+        random_user = str(uuid4())
+        response = fixture_client.get(
+            self.endpoint.format(expid=expid, graph_type="standard", grouped="none"),
+            query_string={"loggedUser": random_user},
+        )
+        resp_obj: dict = response.get_json()
+        assert resp_obj["error_message"] == ""
+        assert resp_obj["error"] == False
+        assert resp_obj["total_jobs"] == len(resp_obj["nodes"])
+
+    def test_graph_standard_datemember_retro3(self, fixture_client: FlaskClient):
+        expid = "a3tb"
+        random_user = str(uuid4())
+        response = fixture_client.get(
+            self.endpoint.format(
+                expid=expid, graph_type="standard", grouped="date-member"
+            ),
+            query_string={"loggedUser": random_user},
+        )
+        resp_obj: dict = response.get_json()
+        assert resp_obj["error_message"] == ""
+        assert resp_obj["error"] == False
+        assert resp_obj["total_jobs"] == len(resp_obj["nodes"])
+
+    def test_graph_standard_status_retro3(self, fixture_client: FlaskClient):
+        expid = "a3tb"
+        random_user = str(uuid4())
+        response = fixture_client.get(
+            self.endpoint.format(expid=expid, graph_type="standard", grouped="status"),
+            query_string={"loggedUser": random_user},
+        )
+        resp_obj: dict = response.get_json()
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert resp_obj["total_jobs"] == len(resp_obj["nodes"])
 
@@ -271,6 +319,7 @@ class TestExpCount:
         response = fixture_client.get(self.endpoint.format(expid=expid))
         resp_obj: dict = response.get_json()
 
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert resp_obj["total"] == sum(
             [resp_obj["counters"][key] for key in resp_obj["counters"]]
@@ -284,6 +333,7 @@ class TestExpCount:
         response = fixture_client.get(self.endpoint.format(expid=expid))
         resp_obj: dict = response.get_json()
 
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert resp_obj["total"] == sum(
             [resp_obj["counters"][key] for key in resp_obj["counters"]]
@@ -308,6 +358,7 @@ class TestSummary:
         )
         resp_obj: dict = response.get_json()
 
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert resp_obj["n_sim"] > 0
 
@@ -322,6 +373,7 @@ class TestStatistics:
         )
         resp_obj: dict = response.get_json()
 
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert resp_obj["Statistics"]["Period"]["From"] == "None"
 
@@ -334,14 +386,11 @@ class TestCurrentConfig:
         response = fixture_client.get(self.endpoint.format(expid=expid))
         resp_obj: dict = response.get_json()
 
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert (
             resp_obj["configuration_filesystem"]["CONFIG"]["AUTOSUBMIT_VERSION"]
             == "4.0.95"
-        )
-        assert (
-            resp_obj["configuration_current_run"]["CONFIG"]["AUTOSUBMIT_VERSION"]
-            == "4.0.101"
         )
 
     def test_retrocomp_v3_conf_files(self, fixture_client: FlaskClient):
@@ -349,6 +398,7 @@ class TestCurrentConfig:
         response = fixture_client.get(self.endpoint.format(expid=expid))
         resp_obj: dict = response.get_json()
 
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert (
             resp_obj["configuration_filesystem"]["conf"]["config"]["AUTOSUBMIT_VERSION"]
@@ -364,6 +414,7 @@ class TestPklInfo:
         response = fixture_client.get(self.endpoint.format(expid=expid, timestamp=0))
         resp_obj: dict = response.get_json()
 
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert len(resp_obj["pkl_content"]) == 8
 
@@ -379,6 +430,7 @@ class TestPklTreeInfo:
         response = fixture_client.get(self.endpoint.format(expid=expid, timestamp=0))
         resp_obj: dict = response.get_json()
 
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert len(resp_obj["pkl_content"]) == 8
 
@@ -394,5 +446,6 @@ class TestExpRunLog:
         response = fixture_client.get(self.endpoint.format(expid=expid))
         resp_obj: dict = response.get_json()
 
+        assert resp_obj["error_message"] == ""
         assert resp_obj["error"] == False
         assert resp_obj["found"] == True